fix two bugs in the rc.subr script
influxd_flags, which usually means command line options to `command`
but when you override `command`, daemon(8) in this case, in the script,
the variable is passed to the command you specified. thus, setting
influxd_flags in rc.conf(5) produces "invalid argument". the correct
variable is influxdb_flags.

it called load_rc_config *after* setting the defaults and expanding the
variables, which effectively *disables* overriding mechanism. thus,
influxdb_flags was never be applied to the command arguments.